+++ change_log.md
@@ -2,8 +2,13 @@
 
 ### bear_export_sync.py
 
+#### Version 1.7.6 - 2018-11-21 at 11:00 IST - Updates:
+- Fix: removing empty elements in CSV in '-t=' and '-x=' values.
+- `bear_import_sync.sh` updated to reflect changes
+
 #### Version 1.7.5 - 2018-11-21 at 07:48 IST - Updates:
-- Fix on issue #12: hardcoded temp vs temp_path. 'BearTemp' is now also renamed to 'BearProc' 
+- Fix on issue #12: hardcoded temp vs temp_path.
+- '~/BearTemp' remamed to '~/BearProc', so '~/BearTemp' and '~/temp' are now obsolete
 
 #### Version 1.7.4 - 2018-11-20 at 21:46 IST - Updates:
 - Fix on issue #18: emtpy arguments confuses rsync! in 'def rsync_files_from_temp()'